+++ rabbit-security-pom/rabbit-security-redisson-cache/src/main/java/com/rabbitframework/security/cache/redisson/RedisManager.java
@@ -4,17 +4,17 @@ import java.util.Set;
 
 public interface RedisManager {
 
-	public byte[] get(String key);
+    public byte[] get(String key);
 
-	public byte[] set(String key, byte[] value);
+    public byte[] set(String key, byte[] value);
 
-	public byte[] set(String key, byte[] value, int expire);
+    public byte[] set(String key, byte[] value, long expire);
 
-	public void del(String key);
+    public void del(String key);
 
-	public void flushDB();
+    public void flushDB();
 
-	public Long dbSize();
+    public Long dbSize();
 
-	public Set<String> keys(String pattern);
+    public Set<String> keys(String pattern);
 }
